Fix animation resources Add Event Bus system

+++ Mario/GameObject.h
@@ -4,7 +4,7 @@ class CGameObject
 	int id;
 
 public:
-	virtual void Update(float dt) = 0;
+	virtual void Update(int dt) = 0;
 	virtual void Render() = 0;
 };
